% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkInteger.R
\name{check.integer}
\alias{check.integer}
\title{Generic function to check integer}
\usage{
check.integer(x)
}
\arguments{
\item{x}{Value to be checked.}
}
\value{
The logical value depending on whether \code{x} is integer or not.
}
\description{
This function checks if a given number n is integer or not. I'm making it internal
by not \code{export}ing.
}
\examples{
check.integer(3)
check.integer(3.25)
}
