% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial.R
\name{fact_func}
\alias{fact_func}
\title{Generic function to get factorial of a number}
\usage{
fact_func(n)
}
\arguments{
\item{n}{the value whose factorial to be taken.}
}
\value{
The \code{n}th factorial is returned.
}
\description{
This function calculates the n factorial of a number. I'm making it external
by \code{export}ing.
}
\examples{
fact_func(6)
fact_func(12)
}
