% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{logarithm}
\alias{logarithm}
\title{Generic Log function}
\usage{
logarithm(x, n = exp(1), plot_it = FALSE)
}
\arguments{
\item{x}{Vector whose log has to be taken.}

\item{n}{Base value for the log.}

\item{plot_it}{Display a plot of \code{x} vs the output? Use logical.}
}
\value{
log of vector \code{x}, to the base of \code{n}.
}
\description{
This function raises a vector to any power. I'm making it internal,
by not \code{export}ing, because I don't want users to use it.
Why not? Because I said so!
}
