% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathfunctions-package.r
\docType{package}
\name{mathfunctions}
\alias{mathfunctions}
\alias{mathfunctions-package}
\title{Convenient Computation of Mathematical Functions}
\description{
Are you tired of using the inbuilt operators, \code{^} or \code{**}
in R or searching for functions to perform common mathematical functions
Use this package to call functions that apply common mathematical operators
to your vectors.
}
\note{
This package isn't actually the first version. There are many inbuilt
functions that does similar things like log. But some functions like nthroot and
nthFibonacci are new.
}
\author{
Arun Rajendran
}
