% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nthFibonacci.R
\name{nthFibonacci}
\alias{nthFibonacci}
\title{Generic function to get nth Fibonacci number}
\usage{
nthFibonacci(n, a = 0, b = 1)
}
\arguments{
\item{n}{the value denoting the position of result to be obtained.}

\item{a}{first term of fibonacci series.}

\item{b}{second term of fibonacci series.}
}
\value{
The \code{n}th fibonacci number of series starting with \code{a},\code{b}.
}
\description{
This function calculates the nth fibonacci number of a series given
first two values. I'm making it external
by \code{export}ing.
}
\examples{
nthFibonacci(6)
nthFibonacci(12,1,2)
}
