% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nthRoot.R
\name{nthroot}
\alias{nthroot}
\title{Generic Root function}
\usage{
nthroot(x, n = 2, plot_it = FALSE)
}
\arguments{
\item{x}{Vector whose root has to be taken.}

\item{n}{Order of the root.}

\item{plot_it}{Display a plot of \code{x} vs the output? Use logical.}
}
\value{
The vector \code{x}, raised to the power of \code{1/n}.
}
\description{
This function calculates the nth root of a vector given n. I'm making it external
by \code{export}ing.
}
\examples{
nthroot(3,2)   #square root
nthroot(-2,3)  #cube root
}
