% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_series.R
\name{sum_series}
\alias{sum_series}
\title{Generic function to get sum of a series}
\usage{
sum_series(a = 1, d = 1, n)
}
\arguments{
\item{a}{the first term.}

\item{d}{difference}

\item{n}{number of terms}
}
\value{
The sum of \code{n} terms of series starting with \code{a} and difference \code{d}.
}
\description{
This function calculates the sum of a finite Arithmetic progression series given first term, the difference
and the number of terms
I'm making it external by \code{export}ing.
}
\examples{
sum_series(1,1,50)
sum_series(12,2,20)
}
